requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
ibm-watson-machine-learning>=1.0.326
llama_index==0.10.9
llama-index-agent-openai==0.1.7
llama-index-core==0.10.33
llama-index-embeddings-elasticsearch==0.1.2
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-llms-watsonx==0.1.7
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-elasticsearch==0.1.7
llama-parse==0.4.2
llamaindex-py-client==0.1.19
elasticsearch==8.13.0
pydantic==2.7.1
ibm-cos-sdk
docx2txt
ibm_cloud_sdk_core
ibm_watson
asyncpg